requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.5
django-cookies-samesite==0.8.0
django-cors-headers==3.7.0
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
mysqlclient==2.0.3
numpy==1.19.5
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pydevd-pycharm==211.7036.9
pykka==3.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
ua-parser==0.10.0
urllib3==1.26.3
whitenoise==5.2.0
xmltodict==0.12.0
https://download.pytorch.org/whl/cpu/torch-1.8.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.9.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
